requirement.txt
aiohttp==3.8.3
aiosignal==1.3.1
appdirs==1.4.4
async-timeout==4.0.2
asynctest==0.13.0
attrs==22.2.0
base58==2.1.1
bitarray==2.6.2
brotlipy==0.7.0
certifi @ file:///croot/certifi_1671487769961/work/certifi
cffi @ file:///croot/cffi_1670423208954/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click==8.1.3
cryptography @ file:///croot/cryptography_1665612644927/work
cycler==0.11.0
cytoolz==0.12.1
docker-pycreds==0.4.0
eth-abi==2.2.0
eth-account==0.5.9
eth-hash==0.5.1
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.2.1
eth-typing==2.3.0
eth-utils==1.9.5
flit_core @ file:///opt/conda/conda-bld/flit-core_1644941570762/work/source/flit_core
fonttools==4.38.0
frozenlist==1.3.3
gitdb==4.0.10
GitPython==3.1.31
h5py==3.7.0
hexbytes==0.3.0
idna @ file:///croot/idna_1666125576474/work
importlib-metadata==6.0.0
importlib-resources==5.10.2
ipfshttpclient==0.8.0a2
jsonschema==4.17.3
kiwisolver==1.4.4
lru-dict==1.1.8
matplotlib==3.5.3
methodtools==0.4.5
mkl-fft==1.3.1
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626179032232/work
mkl-service==2.4.0
more-itertools==9.0.0
multiaddr==0.0.9
multidict==6.0.4
netaddr==0.8.0
numpy @ file:///opt/conda/conda-bld/numpy_and_numpy_base_1653915516269/work
packaging==23.0
params==0.9.0
parsimonious==0.8.1
pathtools==0.1.2
Pillow==9.3.0
pkgutil_resolve_name==1.3.10
protobuf==3.19.5
psutil==5.9.4
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pycryptodome==3.16.0
PyExecJS==1.5.1
pyOpenSSL @ file:///opt/conda/conda-bld/pyopenssl_1643788558760/work
pyparsing==3.0.9
pyrsistent==0.19.3
PySocks @ file:///tmp/build/80754af9/pysocks_1594394576006/work
python-dateutil==2.8.2
python-dotenv==0.21.0
pyvacy==0.0.32
PyYAML==6.0
requests @ file:///opt/conda/conda-bld/requests_1657734628632/work
rlp==2.0.1
scipy==1.7.3
sentry-sdk==1.16.0
setproctitle==1.3.2
six @ file:///tmp/build/80754af9/six_1644875935023/work
smmap==5.0.0
toolz==0.12.0
torch==1.12.1
torchaudio==0.12.1
torchvision==0.13.1
typing_extensions @ file:///croot/typing_extensions_1669924550328/work
urllib3 @ file:///croot/urllib3_1670526988650/work
varint==1.0.2
wandb==0.13.10
web3==5.31.3
websockets==9.1
wirerope==0.4.6
yarl==1.8.2
zipp==3.11.0
